+++ src/components/pages/UsersPage/UsersPage.js
@@ -3,12 +3,12 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faChevronLeft, faChevronRight } from '@fortawesome/free-solid-svg-icons';
 import cx from 'classnames';
 import styled from 'styled-components';
-import UsersTable from './UsersTable';
+import UsersTable from '../../../features/UsersManagement/UsersTable';
 import PageBase from '../PageBase';
 import useDebounce from '../../../hooks/useDebounce';
 import { useUsersQuery } from '../../../service/api';
 import Select from '../../elements/Input/Select';
-import UsersActions from './UsersActions';
+import UsersActions from '../../../features/UsersManagement/UsersActions';
 import { colorGrey } from '../../../styles/colors';
 
 const UsersSection = styled.div`
